llm/requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
boto3==1.37.36
botocore==1.37.36
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
Django==5.2
django-activatable-model==3.1.0
django-entity==6.2.3
django-filter==25.1
django-manager-utils==3.1.5
django-query-builder==3.2.0
djangorestframework==3.16.0
dotenv==0.9.9
drf-nested-routers==0.94.1
drf-yasg==1.21.10
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.18.0
fleming==0.7.0
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langsmith==0.3.32
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.5
orjson==3.10.16
packaging==24.2
peft==0.15.2
pillow==11.2.1
propcache==0.3.1
psutil==7.0.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-utils==0.4.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.5
safetensors==0.5.3
serializers==0.2.4
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
sympy==1.13.1
tenacity==9.1.2
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.1
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
zstandard==0.23.0
